pyproject.toml
[tool.poetry]
name = "data-analysis-system"
version = "0.1.0"
description = ""
authors = ["Suyash Shrestha <sthasuyash11@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
langgraph = "^0.2.67"
langchain = "^0.3.15"
langchain-openai = "^0.3.2"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
e2b-code-interpreter = "^1.0.4"
langchain-community = "^0.3.15"
rich = "^13.9.4"
langchain-google-genai = "^2.0.9"
markdown-pdf = "^1.3.3"
streamlit = "^1.41.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
